python_receiver/requirements.txt
pyserial==3.5
matplotlib==3.10.7
